requirements.txt
django
pytz
psycopg2
djangorestframework
djoser
django_filter